improve error handling

+++ lcbuilder/eleanor_manager.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import shutil
 import traceback
@@ -5,14 +6,19 @@ import sys
 import lcbuilder.eleanor
 sys.modules['eleanor'] = sys.modules['lcbuilder.eleanor']
 import eleanor
-from eleanor.maxsector import maxsector
 
 
 class EleanorManager:
     @staticmethod
     def update():
         eleanor_path = os.path.join(os.path.expanduser('~'), '.eleanor')
-        eleanor.update_max_sector()
+        try:
+            eleanor.update_max_sector()
+        except Exception as e:
+            traceback.print_exc()
+            logging.exception("Can't update ELEANOR max_sector")
+            return
+        from eleanor.maxsector import maxsector
         for sector in range(1, maxsector + 1):
             sectorpath = eleanor_path + '/metadata/s{:04d}'.format(sector)
             if os.path.exists(sectorpath) and os.path.isdir(sectorpath) and not os.listdir(sectorpath):
@@ -22,6 +28,7 @@ class EleanorManager:
                 try:
                     eleanor.Update(sector)
                 except Exception as e:
-                    traceback.print_exc()
-                    shutil.rmtree(sectorpath)
+                    logging.exception("Can't update ELEANOR sector %s", maxsector)
+                    if os.path.exists(sectorpath):
+                        shutil.rmtree(sectorpath)
                     break
